A BIG update...
Major thing is NO-GUI mode is configurable and can persist the counters in a JDBC session. (ASE and H2 has been tested) H2 would be the preferred option. This so we can store the counters in another server/database store than the one we pulling out data from. Using the H2 db it would also be easier to sample ASE counters on Unix systems and later on view them on a Windows machine.

And some changes to "capture SQL" tool.
And saving more options in the "save props" file!
And a lot more small things.