New version of the H2 database is used: 1.2.135 (2010-05-08)<br> This version supports embedded BLOBS in the database, so no extra directories for blobs will be created
Tab: Processes & Active Statements<br>
Added column 'WaitClassDesc', which is the class description for the WaitEventID

Tab: Active Statements<br>
- Added column 'multiSampled', which is 'YES' if the PK (SPID, monSource [,InstanceID]) was visible in the previous sample as well
- Added column 'pssinfo_tempdb_pages' to check how many tempdb pages this SPID is currently using
- Added column 'tempdb_name' to check what tempdb this SIPD is using
- Stacktrace, which is deferred to the localCalculation() method is not executed if the WaitEventID has changed.
- Tooltip on sqlText is not adding extra newlines, which made large outputs a bit unreadable
- Also marked the 'multiSampled' rows with the color 'ORANGE'
- InMemoryHistory: Is now showing the checkboxes and tooltip for Active Statements

Tab: Cached Objects
Changed default postponeTime to 600, every 10 minutes

Tab: Blocking
Is now only available for ase servers over version 15.0.0 ESD#2

Tab: System Load
Rearranged the column layout to make it a bit more readable

Dialog: CheckForNewUpdates
Now doing a better job of remembering the checkbox 'do not show this again'
And some better text descriptions...

Dialog: ConnectionDialog
For the 'offline read', you can start the H2 as a network server

in 'offline read' mode, I added 'Fast Forwad' and 'Rewind' buttons, which goes to the next sample time
that has data for this tab. It's especially good in 'Active Statement' and other tabs where
data is not always present, for example if you have the 'postpone time' set...
The Buttons are located in every tab in the 'Sample Information' panel.

Added a "rowcount" label to indicate how many rows there are in the JTable.
It's located "Show Counter Type" panel

Some changes how data is displayed in the 'offline read' if data can't be found for the selected sample period

Connection dialog:
in tab 'Counter Storage', Changed the layout of the panel/JTable "What Counter Data should be Persisted"

Offline wizard:
in page 3 'What should be offline sampled', Changed the layout of the panel/JTable
And added the button "Set to Template", so we can reuse what we are currently doing in the GUI.

Upgraded to latest H2 version, 1.2.137 (2010-06-06)